1-array_iterator.c

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
-#include "function_pointers.h"
 #include <stdlib.h>
+#include "function_pointers.h"
 
 /**
  * array_iterator - iterates through array
@@ -11,7 +11,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int i;
+	size_t i;
 
 	if (array == NULL || action == NULL)
 		return;
